+++ apps/app/src/test/scala/com/daml/network/integration/tests/DecentralizedSynchronizerMigrationIntegrationTest.scala
@@ -725,7 +725,7 @@ class DecentralizedSynchronizerMigrationIntegrationTest
               }
 
             withClueAndLog("migrate the late joining node") {
-              // sv1 is the founder so specifically join it later to validate our replay
+              // sv1 is the sv1 so specifically join it later to validate our replay
               sv1LocalBackend.startSync()
 
               val changeSerial = namespaceChangeResult.map(_.base.serial).max
@@ -906,7 +906,7 @@ class DecentralizedSynchronizerMigrationIntegrationTest
                     new DsoRules_AddSv(
                       "bob",
                       "Bob",
-                      SvUtil.DefaultFoundingNodeWeight,
+                      SvUtil.DefaultSV1Weight,
                       "bob-participant-id",
                       new Round(42),
                     )