+++ apps/app/src/test/scala/com/daml/network/integration/tests/DecentralizedSynchronizerSoftDomainMigrationIntegrationTest.scala
@@ -233,7 +233,7 @@ class DecentralizedSynchronizerSoftDomainMigrationIntegrationTest
               new dsorules.DsoRules_AddSv(
                 "alice",
                 "Alice",
-                SvUtil.DefaultFoundingNodeWeight,
+                SvUtil.DefaultSV1Weight,
                 "alice-participant-id",
                 dummyRound,
               )
@@ -301,7 +301,7 @@ class DecentralizedSynchronizerSoftDomainMigrationIntegrationTest
         new so.SvOnboardingConfirmed(
           sv1Party.toProtoPrimitive,
           "irrelevant name",
-          SvUtil.DefaultFoundingNodeWeight,
+          SvUtil.DefaultSV1Weight,
           "PAR::sv::1220f3e2",
           "observing domain migration",
           dsoParty.toProtoPrimitive,