+++ apps/scan/src/main/scala/com/daml/network/scan/store/db/DbScanStore.scala
@@ -62,7 +62,7 @@ object DbScanStore {
 class DbScanStore(
     override val key: ScanStore.Key,
     storage: DbStorage,
-    isFounder: Boolean,
+    isFirstSv: Boolean,
     override protected val loggerFactory: NamedLoggerFactory,
     override protected val retryProvider: RetryProvider,
     createScanAggregatesReader: DbScanStore => ScanAggregatesReader,
@@ -127,7 +127,7 @@ class DbScanStore(
       new ScanAggregator(
         storage,
         storeId,
-        isFounder,
+        isFirstSv,
         createScanAggregatesReader(this),
         loggerFactory,
         domainMigrationId,